refactor(generator): split ModuleRecords out of DiscoveryData
